RenderingPanel
Started the RenderingPanel class. Has been set up so that it can take an
Area to render, as well as overriding the paintComponent method so that
it can draw that Area from one direction. I've also made small changes
to the Area and Main classes, so that we can temporary import data from
standard text files.