Tiles Map
Created a method in Main which reads a file containing information on
the tiles in the game. Parsing the file results in a mapping of integers
to tiles, where the integer values are what is stored in the files
containing Area data. At the moment this method reads from a standard
text file, this will need to be changed to an xml file when that has
been implemented.