Rendering Panel Updates (drawing / recognising items)
Updated the rendering panel further so that it now draws items that are
contained in the area, as well as creates a bounding box for them. This
means that they can be recognised when they are clicked on. The creation
of the bounding boxes seems inefficient and could be streamlined at a
later date.