Comparing Item Positions
Added a comparator to the RenderingPanel class that organises Items in
the Area based on their position. This way Items that are 'closer' are
positioned earlier in Items list. This will help in determining what
Item was actually clicked on if two items are overlapping. Currently
this only works for facing north, I will add more comparators for the
other directions after this one has been tested.