Added Menu Bar and Directions
There is now a menu bar to replace the lone exit button that we had
previously. There are now direction variables in GameFrame which are
used to describe the direction the user is facing. Using the WASD keys
the user can either rotate the direction (using A and D) or set the
direction to north or south(W and S respectively).