Drawing From Four Directions
I've completed an initial implementation of drawing from four different
directions. I've also adjusted the Tile classes to factor this in. Now
rather than taking an image as a parameter, a Tile takes the base
filename and creates an array of images by attaching North, South, East
and West to the end of the filename.