Bounding Boxes and Rotation Working!
Rotation is now working correctly, with all tiles being drawn in the
correct position when rotated. The points also rotate correctly, which
was accomplished by creating a class called BoundingBox which is an
extension of Polygon. The only difference is that BoundingBox also takes
a point as a parameter, which relates to the position that it is
representing.